moving much of the logic to nl_interface and leaving only the NLP specific bits in the Model class in nl.py
